big fux

+++ src/database/migrations/2024_03_30_130726_create_character_contest_table.php
@@ -19,8 +19,8 @@ return new class extends Migration
             $table->foreignId('contest_id')->onDelete('cascade');
             $table->foreignId('enemy_id')->onDelete('cascade');
 
-            $table->integer('hero_hp');
-            $table->integer('enemy_hp');
+            $table->float('hero_hp');
+            $table->float('enemy_hp');
             $table->softDeletes();
         });
     }
